backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
requests==2.31.0
google-generativeai==0.8.4
bcrypt==4.2.0
# ML / Handwriting Analysis
tensorflow>=2.16.0
Pillow>=10.0.0
numpy>=1.26.0
